Give Timestamp its own internal object so it stays distinct from Date
Commit 2 had to scatter Timestamp's derived accessors into package-level
extensions (and re-export them) because, sharing `object internal` with
`Date`/`Year`/`Day`, their identically-named accessors collided after
erasure and were ambiguous wherever the opaque types were transparent.

Put `Timestamp` in its own `object timestampInternal` instead. Now `Date`
and `Timestamp` are *distinct* opaque types everywhere except each one's own
defining object, so all of Timestamp's accessors (`date`/`time`/`year`/
`month`/`day`/`monthstamp`/`hour`/`minute`/`second`/`in`/`stdlib`/`instant`)
plus its Showable/Decodable givens live in its own companion — no
package-level extensions and no extra `soundness` exports. `Timestamp` is
built from `Date` purely through the public `date.jdn` / `Date.julianDay`
API; the shared grid constant `MillisPerDay` is now `private[aviation]`.

All 427 aviation tests pass.
